requirements.txt
twine
grpcio==1.24.1
grpcio-tools==1.24.1
bottle
Paste
wheel
bottle-cors